refactor: 영화 목록 렌더 함수 분리
